move DashboardMetadataDAO interface to internal.database package
